move endpoint into extended trs api
