+++ dockstore-webservice/src/main/resources/openapi3/openapi.yaml
@@ -491,6 +491,40 @@ paths:
       summary: List entries of an organization
       tags:
       - extendedGA4GH
+  /api/ga4gh/v2/extended/entries:
+    get:
+      description: Retrieve an abbreviated description and the default version name
+        of each published entry.
+      operationId: getAllEntries
+      parameters:
+      - description: Pagination offset
+        in: query
+        name: offset
+        schema:
+          type: integer
+          format: int32
+          default: 0
+          minimum: 0
+      - description: Pagination limit
+        in: query
+        name: limit
+        schema:
+          type: integer
+          format: int32
+          default: 100
+          maximum: 100
+          minimum: 1
+      responses:
+        "200":
+          content:
+            application/json:
+              schema:
+                type: array
+                items:
+                  $ref: '#/components/schemas/EntryLiteAndVersionName'
+          description: Successfully retrieved entries
+      tags:
+      - extendedGA4GH
   /api/ga4gh/v2/extended/entryVersionsToAggregate:
     get:
       description: This endpoint gets entry versions that have new execution metrics
@@ -3316,40 +3350,6 @@ paths:
       - BEARER: []
       tags:
       - curation
-  /entries:
-    get:
-      description: Retrieve an abbreviated description and the default version name
-        of each published entry.
-      operationId: getAllEntries
-      parameters:
-      - description: Pagination offset
-        in: query
-        name: offset
-        schema:
-          type: integer
-          format: int32
-          default: 0
-          minimum: 0
-      - description: Pagination limit
-        in: query
-        name: limit
-        schema:
-          type: integer
-          format: int32
-          default: 100
-          maximum: 100
-          minimum: 1
-      responses:
-        "200":
-          content:
-            application/json:
-              schema:
-                type: array
-                items:
-                  $ref: '#/components/schemas/EntryLiteAndVersionName'
-          description: Successfully retrieved entries
-      tags:
-      - entries
   /entries/toCategorize:
     get:
       description: "Get published entries that are new and uncategorized, or that\
